requirements.txt
dlib==19.24.1
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.21.0
opencv-python==4.5.3.56
Pillow==8.3.1 